1、升级spring boot版本为2.6.2 2、升级mybatis-plus版本为3.5.0 3、升级knife4j为3.0版本

+++ jeeagile-frame/jeeagile-frame-api/src/main/java/com/jeeagile/frame/service/IAgileBaseTreeService.java
@@ -22,5 +22,5 @@ public interface IAgileBaseTreeService<T> extends IAgileBaseService<T> {
      * @param parentUuid
      * @return
      */
-    int countChild(Serializable parentUuid);
+    long countChild(Serializable parentUuid);
 }
